requirements.txt
jax==0.6.0
jaxlib>=0.6.0
matplotlib>=3.5
numpy>=1.21
pyyaml